Gunnar: fix multiple runtime issues         - usage of reorthogonalized projected Lanczos in QHEOnSphereMainTask works fine         - uninitialized value in GenericSignalHandler fixed
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@1797 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ src/Options/SingleIntegerOption.h
@@ -44,7 +44,7 @@ class SingleIntegerOption : public AbstractOption
   long Integer;
   
   // default value
-  double DefaultValue;
+  long DefaultValue;
 
   // maximum value allowed
   long MaxValue;
